fix @domoritz's comments and fixes #128
