--- src/vegalite.js
@@ -1,1603 +1,0 @@
-(function(root, factory) {
-  if (typeof define === 'function' && define.amd) {
-    // AMD. Register as an anonymous module.
-    define([], factory);
-  } else if (typeof exports === 'object') {
-    // Node. Does not work with strict CommonJS, but
-    // only CommonJS-like environments that support module.exports,
-    // like Node.
-    module.exports = factory();
-  } else {
-    // Browser globals (root is window)
-    root.vl = factory();
-  }
-}(this, function() {
-
-// BEGINNING OF THIS MODULE
-
-var vl = {};
-var TABLE = "table";
-var STACKED = "stacked";
-var INDEX = "index";
-
-var X = "x";
-var Y = "y";
-var ROW = "row";
-var COL = "col";
-var SIZE = "size";
-var SHAPE = "shape";
-var COLOR = "color";
-var ALPHA = "alpha";
-var TEXT = "text";
-
-var O = 1;
-var Q = 2;
-var T = 4;
-
-
-var consts = {};
-
-
-consts.encodings = {X:X, Y:Y, ROW:ROW, COL:COL, SIZE:SIZE, SHAPE:SHAPE, COLOR:COLOR, ALPHA:ALPHA, TEXT:TEXT};
-consts.dataTypes = {"O": O, "Q": Q, "T": T};
-
-// inverse mapping e.g., 1=>O
-consts.dataTypeNames = ["O","Q","T"].reduce(function(r,x) {
-  r[consts.dataTypes[x]] = x; return r;
-},{});
-
-consts.DEFAULTS = {
-  // template
-  width: undefined,
-  height: undefined,
-  viewport: undefined,
-  _minWidth: 20,
-  _minHeight: 20,
-
-  // data source
-  dataUrl: undefined, //for easier export
-  useVegaServer: false,
-  vegaServerUrl: "http://localhost:3001",
-  vegaServerTable: undefined,
-  dataFormatType: "json",
-
-  //small multiples
-  cellHeight: 200, // will be overwritten by bandWidth
-  cellWidth: 200, // will be overwritten by bandWidth
-  cellPadding: 0.1,
-  cellBackgroundColor: "#fdfdfd",
-  xAxisMargin: 80, //TODO rename
-  yAxisMargin: 0,//TODO rename
-  textCellWidth: 90,
-
-  // marks
-  bandSize: 21,
-  bandPadding: 1,
-  pointSize: 50,
-  pointShape: "circle",
-  strokeWidth: 2,
-  color: "steelblue",
-  textColor: "black",
-  textAlign: "left",
-  textBaseline: "middle",
-  textMargin: 4,
-  font: "Helvetica Neue",
-  fontSize: "12",
-  fontWeight: "normal",
-  fontStyle: "normal",
-  opacity: 1,
-  _thickOpacity: 0.5,
-  _thinOpacity: 0.2,
-
-  // scales
-  // TODO remove _xZero, ...
-  _xZero: true,
-  _xReverse: false,
-  _yZero: true,
-  _yReverse: false,
-  timeScaleNice: "day"
-};
-
-
-var schema = {};
-schema.aggr = {
-  type: "string",
-  enum: ["avg", "sum", "min", "max", "count"],
-  supportedEnums: {
-    Q: ["avg", "sum", "min", "max", "count"],
-    O: ["count"],
-    T: ["avg", "min", "max", "count"],
-    "": ["count"],
-  },
-  supportedTypes: {"Q": true, "O": true, "T": true, "": true}
-};
-
-schema.timefns = ["month", "year", "day", "date", "hour", "minute", "second"];
-
-// vl.schema.aggr.enum
-schema.aggr.enum = ["avg", "sum", "min", "max", "count"];
-
-// vl.schema.timefns
-schema.timefns = ["month", "year", "day", "date", "hour", "minute", "second"];
-// vl.schema.scale_type.enum
-schema.scale_type = {};
-schema.scale_type.enum = ["-", "log","pow", "sqrt", "quantile"];
-
-
-var MAX_BINS = 20;
-
-var util = {};
-
-util.keys = function (obj) {
-  var k = [], x;
-  for (x in obj) k.push(x);
-  return k;
-}
-
-util.vals = function (obj) {
-  var v = [], x;
-  for (x in obj) v.push(obj[x]);
-  return v;
-}
-
-util.range = function (start, stop, step) {
-  if (arguments.length < 3) {
-    step = 1;
-    if (arguments.length < 2) {
-      stop = start;
-      start = 0;
-    }
-  }
-  if ((stop - start) / step == Infinity) throw new Error("infinite range");
-  var range = [], i = -1, j;
-  if (step < 0) while ((j = start + step * ++i) > stop) range.push(j);
-  else while ((j = start + step * ++i) < stop) range.push(j);
-  return range;
-}
-
-util.find = function (list, pattern) {
-  var l = list.filter(function(x) {
-    return x[pattern.name] === pattern.value;
-  });
-  return l.length && l[0] || null;
-}
-
-util.uniq = function (data, field) {
-  var map = {}, count = 0, i, k;
-  for (i=0; i<data.length; ++i) {
-    k = data[i][field];
-    if (!map[k]) {
-      map[k] = 1;
-      count += 1;
-    }
-  }
-  return count;
-}
-
-util.minmax = function (data, field) {
-  var stats = {min: +Infinity, max: -Infinity};
-  for (i=0; i<data.length; ++i) {
-    var v = data[i][field];
-    if (v > stats.max) stats.max = v;
-    if (v < stats.min) stats.min = v;
-  }
-  return stats;
-}
-
-util.duplicate = function (obj) {
-  return JSON.parse(JSON.stringify(obj));
-};
-
-util.any = function(arr, f){
-  var i=0, k;
-  for (k in arr) {
-    if(f(arr[k], k, i++)) return true;
-  }
-  return false;
-}
-
-util.all = function(arr, f){
-  var i=0, k;
-  for (k in arr) {
-    if(!f(arr[k], k, i++)) return false;
-  }
-  return true;
-}
-
-util.merge = function(dest, src){
-  return util.keys(src).reduce(function(c, k){
-    c[k] = src[k];
-    return c;
-  }, dest);
-};
-
-util.getbins = function (stats) {
-  return vg.bins({
-    min: stats.min,
-    max: stats.max,
-    maxbins: MAX_BINS
-  });
-}
-
-util.error = function(msg){
-  console.error("[VL Error]", msg);
-}
-
-// ----
-vl.Encoding = (function() {
-
-  function Encoding(marktype, enc, config) {
-    this._marktype = marktype;
-    this._enc = enc; // {encType1:field1, ...}
-    this._cfg = util.merge(Object.create(consts.DEFAULTS), config);
-  }
-
-  var proto = Encoding.prototype;
-
-  proto.marktype = function() {
-    return this._marktype;
-  };
-
-  proto.is = function(m) {
-    return this._marktype === m;
-  };
-
-  proto.has = function(x) {
-    return this._enc[x] !== undefined;
-  };
-
-  proto.enc = function(x){
-    return this._enc[x];
-  };
-
-  // get "field" property for vega
-  proto.field = function(x, nodata, nofn) {
-    if (!this.has(x)) return null;
-
-    var f = (nodata ? "" : "data.");
-
-    if (this._enc[x].aggr === "count") {
-      return f + "count";
-    } else if (!nofn && this._enc[x].bin) {
-      return f + "bin_" + this._enc[x].name;
-    } else if (!nofn && this._enc[x].aggr) {
-      return f + this._enc[x].aggr + "_" + this._enc[x].name;
-    } else if (!nofn && this._enc[x].fn){
-      return f + this._enc[x].fn + "_" + this._enc[x].name;
-    } else {
-      return f + this._enc[x].name;
-    }
-  };
-
-  proto.fieldName = function(x){
-    return this._enc[x].name;
-  }
-
-  proto.scale = function(x){
-    return this._enc[x].scale || {};
-  }
-
-  proto.axis = function(x){
-    return this._enc[x].axis || {};
-  }
-
-  proto.aggr = function(x){
-    return this._enc[x].aggr;
-  }
-
-  proto.bin = function(x){
-    return this._enc[x].bin;
-  }
-
-  proto.legend = function(x){
-    return this._enc[x].legend;
-  }
-
-  proto.fn = function(x){
-    return this._enc[x].fn;
-  }
-
-  proto.any = function(f){
-    return util.any(this._enc, f);
-  }
-
-  proto.all = function(f){
-    return util.all(this._enc, f);
-  }
-
-  proto.length = function(){
-    return util.keys(this._enc).length;
-  }
-
-  proto.reduce = function(f, init){
-    var r = init, i=0;
-    for (k in this._enc){
-      r = f(r, this._enc[k], k, this._enc);
-    }
-    return r;
-  }
-
-  proto.forEach = function(f) {
-    var i=0, k;
-    for (k in this._enc) {
-      f(k, this._enc[k], i++);
-    }
-  };
-
-  proto.type = function(x) {
-    return this.has(x) ? this._enc[x].type : null;
-  };
-
-  proto.isType = function(x, t) {
-    var xt = this.type(x);
-    if (xt == null) return false;
-    return (xt & t) > 0;
-  };
-
-  proto.config = function(name) {
-    return this._cfg[name];
-  };
-
-  proto.toSpec = function(excludeConfig){
-    var enc = util.duplicate(this._enc),
-      spec;
-
-    // convert type's bitcode to type name
-    for(var e in enc){
-      enc[e].type = consts.dataTypeNames[enc[e].type];
-    }
-
-    spec = {
-      marktype: this._marktype,
-      enc: enc
-    }
-
-    if(!excludeConfig){
-      spec.cfg = util.duplicate(this._cfg)
-    }
-
-    return spec;
-  };
-
-  proto.toShorthand = function(){
-    var enc = this._enc;
-    return this._marktype + "." + util.keys(enc).map(function(e){
-      var v = enc[e];
-        return e + "-" +
-          (v.aggr ? v.aggr+"_" : "") +
-          (v.fn ? v.fn+"_" : "") +
-          (v.bin ? "bin_" : "") +
-          (v.name || "") + "-" +
-          consts.dataTypeNames[v.type];
-      }
-    ).join(".");
-  }
-
-  Encoding.parseShorthand = function(shorthand, cfg){
-    var enc = shorthand.split("."),
-      marktype = enc.shift();
-
-    enc = enc.reduce(function(m, e){
-      var split = e.split("-"),
-        enctype = split[0],
-        o = {name: split[1], type: consts.dataTypes[split[2]]};
-
-      // check aggregate type
-      for(var i in schema.aggr.enum){
-        var a = schema.aggr.enum[i];
-        if(o.name.indexOf(a+"_") == 0){
-          o.name = o.name.substr(a.length+1);
-          if (a=="count" && o.name.length === 0) o.name = "*";
-          o.aggr = a;
-          break;
-        }
-      }
-      // check time fn
-      for(var i in schema.timefns){
-        var f = schema.timefns[i];
-        if(o.name && o.name.indexOf(f+"_") == 0){
-          o.name = o.name.substr(o.length+1);
-          o.fn = f;
-          break;
-        }
-      }
-
-      // check bin
-      if(o.name && o.name.indexOf("bin_") == 0){
-        o.name = o.name.substr(4);
-        o.bin = true;
-      }
-
-      m[enctype] = o;
-      return m;
-    }, {});
-
-    return new Encoding(marktype, enc, cfg);
-  }
-
-  Encoding.fromSpec = function(spec, extraCfg) {
-    var enc = util.duplicate(spec.enc);
-
-    //convert type from string to bitcode (e.g, O=1)
-    for(var e in enc){
-      enc[e].type = consts.dataTypes[enc[e].type];
-    }
-
-    return new Encoding(spec.marktype, enc, util.merge(spec.cfg, extraCfg || {}));
-  }
-
-  return Encoding;
-
-})();
-
-// ----
-
-vl.getDataUrl = function getDataUrl(encoding, stats) {
-  if (!encoding.config("useVegaServer")) {
-    // don't use vega server
-    return encoding.config("dataUrl");
-  }
-
-  if (encoding.length() === 0) {
-    // no fields
-    return;
-  }
-
-  var fields = []
-  encoding.forEach(function(encType, field){
-    var obj = {
-      name: encoding.field(encType, true),
-      field: field.name
-    }
-    if (field.aggr) {
-      obj.aggr = field.aggr
-    }
-    if (field.bin) {
-      obj.binSize = util.getbins(stats[field.name]).step;
-    }
-    fields.push(obj);
-  });
-
-  var query = {
-    table: encoding.config("vegaServerTable"),
-    fields: fields
-  }
-
-  return encoding.config("vegaServerUrl") + "/query/?q=" + JSON.stringify(query)
-}
-
-// Returns the stats for the dataset.
-// Stats is a map from each field name to an object with min, max, count, cardinality and type.
-vl.getStats = function(data){ // hack
-  var stats = {};
-  var fields = util.keys(data[0]);
-
-  fields.forEach(function(k) {
-    var stat = util.minmax(data, k);
-    stat.cardinality = util.uniq(data, k);
-    //TODO(kanitw): better type inference here
-    stat.type = (typeof data[0][k] === "number") ? "Q" :
-      isNaN(Date.parse(data[0][k])) ? "O" : "T";
-    stat.count = data.length;
-    stats[k] = stat;
-  });
-  return stats;
-}
-
-function getCardinality(encoding, encType, stats){
-  var field = encoding.fieldName(encType);
-  if (encoding.bin(encType)) {
-    var bins = util.getbins(stats[field]);
-    return (bins.stop - bins.start) / bins.step;
-  }
-  return stats[field].cardinality;
-}
-
-function setSize(encoding, stats) {
-  var hasRow = encoding.has(ROW),
-      hasCol = encoding.has(COL),
-      hasX = encoding.has(X),
-      hasY = encoding.has(Y);
-
-  // HACK to set chart size
-  // NOTE: this fails for plots driven by derived values (e.g., aggregates)
-  // One solution is to update Vega to support auto-sizing
-  // In the meantime, auto-padding (mostly) does the trick
-  //
-  var colCardinality = hasCol ? getCardinality(encoding, COL, stats) : 1,
-    rowCardinality = hasRow ? getCardinality(encoding, ROW, stats) : 1;
-
-  var cellWidth = hasX ?
-      +encoding.config("cellWidth") || encoding.config("width") * 1.0 / colCardinality :
-      encoding.marktype() === "text" ?
-        +encoding.config("textCellWidth") :
-        +encoding.config("bandSize"),
-    cellHeight = hasY ?
-      +encoding.config("cellHeight") || encoding.config("height") * 1.0 / rowCardinality :
-      +encoding.config("bandSize"),
-    cellPadding = encoding.config("cellPadding"),
-    bandPadding = encoding.config("bandPadding"),
-    width = encoding.config("_minWidth"),
-    height = encoding.config("_minHeight");
-
-  if (hasX && (encoding.isType(X, O) || encoding.bin(X))) { //ordinal field will override parent
-    // bands within cell use rangePoints()
-    var xCardinality = getCardinality(encoding, X, stats);
-    cellWidth = (xCardinality + bandPadding) * +encoding.config("bandSize");
-  }
-  // Cell bands use rangeBands(). There are n-1 padding.  Outerpadding = 0 for cells
-  width = cellWidth * ((1 + cellPadding) * (colCardinality-1) + 1);
-
-  if (hasY && (encoding.isType(Y, O) || encoding.bin(Y))) {
-    // bands within cell use rangePoint()
-    var yCardinality = getCardinality(encoding, Y, stats);
-    cellHeight = (yCardinality + bandPadding) * +encoding.config("bandSize");
-  }
-  // Cell bands use rangeBands(). There are n-1 padding.  Outerpadding = 0 for cells
-  height = cellHeight * ((1 + cellPadding) * (rowCardinality-1) + 1);
-
-  return {
-    cellWidth: cellWidth,
-    cellHeight: cellHeight,
-    width: width,
-    height:height
-  };
-}
-
-vl.toVegaSpec = function(encoding, stats) {
-  var size = setSize(encoding, stats),
-    cellWidth = size.cellWidth,
-    cellHeight = size.cellHeight;
-
-  var hasAgg = encoding.any(function(v, k){
-    return v.aggr !== undefined;
-  });
-
-  var spec = template(encoding, size, stats),
-    group = spec.marks[0],
-    mark = marks[encoding.marktype()],
-    mdef = markdef(mark, encoding, {
-      hasAggregate: hasAgg
-    });
-
-  var hasRow = encoding.has(ROW), hasCol = encoding.has(COL);
-
-  var preaggregatedData = encoding.config("useVegaServer");
-
-  group.marks.push(mdef);
-  // TODO: return value not used
-  binning(spec.data[0], encoding, {preaggregatedData: preaggregatedData});
-
-  var lineType = marks[encoding.marktype()].line;
-
-  if(!preaggregatedData){
-    encoding.forEach(function(encType, field){
-      if(field.type === T && field.fn){
-        timeTransform(spec.data[0], encoding, encType, field);
-      }
-    });
-  }
-
-  // handle subfacets
-  var aggResult = aggregates(spec.data[0], encoding, {preaggregatedData: preaggregatedData}),
-    details = aggResult.details,
-    hasDetails = details && details.length > 0,
-    stack = hasDetails && stacking(spec, encoding, mdef, aggResult.facets);
-
-  if (hasDetails && (stack || lineType)) {
-    //subfacet to group stack / line together in one group
-    subfacet(group, mdef, details, stack, encoding);
-  }
-
-  // auto-sort line/area values
-  //TODO(kanitw): have some config to turn off auto-sort for line (for line chart that encodes temporal information)
-  if (lineType) {
-    var f = (encoding.isType(X, Q | T) && encoding.isType(Y, O)) ? Y : X;
-    if (!mdef.from) mdef.from = {};
-    mdef.from.transform = [{type: "sort", by: encoding.field(f)}];
-  }
-
-  // Small Multiples
-  if (hasRow || hasCol) {
-    spec = facet(group, encoding, cellHeight, cellWidth, spec, mdef, stack, stats);
-  } else {
-    group.scales = scale.defs(scale.names(mdef.properties.update), encoding,
-      {stack: stack, stats: stats});
-    group.axes = axis.defs(axis.names(mdef.properties.update), encoding);
-    group.legends = legends.defs(encoding);
-  }
-
-  return spec;
-}
-
-function facet(group, encoding, cellHeight, cellWidth, spec, mdef, stack, stats) {
-    var enter = group.properties.enter;
-    var facetKeys = [], cellAxes = [];
-
-    var hasRow = encoding.has(ROW), hasCol = encoding.has(COL);
-
-    var xAxisMargin = encoding.has(Y) ? encoding.config("xAxisMargin") : undefined;
-
-    enter.fill = {value: encoding.config("cellBackgroundColor")};
-
-    //move "from" to cell level and add facet transform
-    group.from = {data: group.marks[0].from.data};
-
-    if (group.marks[0].from.transform) {
-      delete group.marks[0].from.data; //need to keep transform for subfacetting case
-    } else {
-      delete group.marks[0].from;
-    }
-    if (hasRow) {
-      if (!encoding.isType(ROW, O)) {
-        util.error("Row encoding should be ordinal.");
-      }
-      enter.y = {scale: ROW, field: "keys." + facetKeys.length};
-      enter.height = {"value": cellHeight}; // HACK
-
-      facetKeys.push(encoding.field(ROW));
-
-      var from;
-      if (hasCol) {
-        from = util.duplicate(group.from);
-        from.transform = from.transform || [];
-        from.transform.unshift({type: "facet", keys: [encoding.field(COL)]});
-      }
-
-      var axesGrp = groupdef("x-axes", {
-          axes: encoding.has(X) ?  axis.defs(["x"], encoding) : undefined,
-          x: hasCol ? {scale: COL, field: "keys.0", offset: xAxisMargin} : {value: xAxisMargin},
-          width: hasCol && {"value": cellWidth}, //HACK?
-          from: from
-        });
-
-      spec.marks.push(axesGrp);
-      (spec.axes = spec.axes || [])
-      spec.axes.push.apply(spec.axes, axis.defs(["row"], encoding));
-    } else { // doesn't have row
-      if(encoding.has(X)){
-        //keep x axis in the cell
-        cellAxes.push.apply(cellAxes, axis.defs(["x"], encoding));
-      }
-    }
-
-    if (hasCol) {
-      if (!encoding.isType(COL, O)) {
-        util.error("Col encoding should be ordinal.");
-      }
-      enter.x = {scale: COL, field: "keys." + facetKeys.length};
-      enter.width = {"value": cellWidth}; // HACK
-
-      facetKeys.push(encoding.field(COL));
-
-      var from;
-      if (hasRow) {
-        from = util.duplicate(group.from);
-        from.transform = from.transform || [];
-        from.transform.unshift({type: "facet", keys: [encoding.field(ROW)]});
-      }
-
-      var axesGrp = groupdef("y-axes", {
-        axes: encoding.has(Y) ? axis.defs(["y"], encoding) : undefined,
-        y: hasRow && {scale: ROW, field: "keys.0"},
-        x: hasRow && {value: xAxisMargin},
-        height: hasRow && {"value": cellHeight}, //HACK?
-        from: from
-      });
-
-      spec.marks.push(axesGrp);
-      (spec.axes = spec.axes || [])
-      spec.axes.push.apply(spec.axes, axis.defs(["col"], encoding, {
-        xAxisMargin: xAxisMargin
-      }));
-    } else { // doesn't have col
-      if(encoding.has(Y)){
-        cellAxes.push.apply(cellAxes, axis.defs(["y"], encoding));
-      }
-    }
-
-    if(hasRow){
-      if(enter.x) enter.x.offset= xAxisMargin;
-      else enter.x = {value: xAxisMargin};
-    }
-    if(hasCol){
-      //TODO fill here..
-    }
-
-    // assuming equal cellWidth here
-    // TODO: support heterogenous cellWidth (maybe by using multiple scales?)
-    spec.scales = scale.defs(
-      scale.names(enter).concat(scale.names(mdef.properties.update)),
-      encoding,
-      {cellWidth: cellWidth, cellHeight: cellHeight, stack: stack, facet:true, stats: stats}
-    ); // row/col scales + cell scales
-
-    if (cellAxes.length > 0) {
-      group.axes = cellAxes;
-    }
-
-    // add facet transform
-    var trans = (group.from.transform || (group.from.transform = []));
-    trans.unshift({type: "facet", keys: facetKeys});
-
-  return spec;
-  }
-
-function subfacet(group, mdef, details, stack, encoding) {
-  var m = group.marks,
-    g = groupdef("subfacet", {marks: m});
-
-  group.marks = [g];
-  g.from = mdef.from;
-  delete mdef.from;
-
-  //TODO test LOD -- we should support stack / line without color (LOD) field
-  var trans = (g.from.transform || (g.from.transform = []));
-  trans.unshift({type: "facet", keys: details});
-
-  if (stack && encoding.has(COLOR)) {
-    trans.unshift({type: "sort", by: encoding.field(COLOR)});
-  }
-}
-
-function getTimeFn(fn){
-  switch(fn){
-    case "second": return "getUTCSeconds";
-    case "minute": return "getUTCMinutes";
-    case "hour": return "getUTCHours";
-    case "day": return "getUTCDay";
-    case "date": return "getUTCDate";
-    case "month": return "getUTCMonth";
-    case "year": return "getUTCFullYear";
-  }
-  console.error("no function specified for date");
-}
-
-function timeTransform(spec, encoding, encType, field){
-  var func = getTimeFn(field.fn);
-
-  spec.transform = spec.transform || [];
-  spec.transform.push({
-    type: "formula",
-    field: encoding.field(encType),
-    expr: "new Date(d.data."+field.name+")."+func+"()"
-  });
-  return spec;
-}
-
-function binning(spec, encoding, opt) {
-  opt = opt || {};
-  var bins = {};
-  encoding.forEach(function(vv, d) {
-    if (d.bin) bins[d.name] = d.name;
-  });
-  bins = util.keys(bins);
-
-  if (bins.length === 0 || opt.preaggregatedData) return false;
-
-  if (!spec.transform) spec.transform = [];
-  bins.forEach(function(d) {
-    spec.transform.push({
-      type: "bin",
-      field: "data." + d,
-      output: "data.bin_" + d,
-      maxbins: MAX_BINS
-    });
-  });
-  return bins;
-}
-
-function aggregates(spec, encoding, opt) {
-  opt = opt || {};
-  var dims = {}, meas = {}, detail = {}, facets={};
-  encoding.forEach(function(encType, field) {
-    if (field.aggr) {
-      if(field.aggr==="count"){
-        meas["count"] = {op:"count", field:"*"};
-      }else{
-        meas[field.aggr+"|"+field.name] = {
-          op:field.aggr,
-          field:"data."+field.name
-        };
-      }
-    } else {
-      dims[field.name] = encoding.field(encType);
-      if (encType==ROW || encType == COL){
-        facets[field.name] = dims[field.name];
-      }else if (encType !== X && encType !== Y) {
-        detail[field.name] = dims[field.name];
-      }
-    }
-  });
-  dims = util.vals(dims);
-  meas = util.vals(meas);
-
-  if (meas.length > 0 && !opt.preaggregatedData) {
-    if (!spec.transform) spec.transform = [];
-    spec.transform.push({
-      type: "aggregate",
-      groupby: dims,
-      fields: meas
-    });
-
-    if (encoding.marktype() === TEXT) {
-      meas.forEach( function (m) {
-        var fieldName = m.field.substr(5), //remove "data."
-          field = "data." + (m.op ? m.op + "_" : "") + fieldName;
-        spec.transform.push({
-          type: "formula",
-          field: field,
-          expr: "d3.format('.2f')(d."+field+")"
-        });
-      });
-    }
-  }
-  return {
-    details: util.vals(detail),
-    dims: dims,
-    facets: util.vals(facets),
-    aggregated: meas.length > 0
-  }
-}
-
-function stacking(spec, encoding, mdef, facets) {
-  if (!marks[encoding.marktype()].stack) return false;
-  if (!encoding.has(COLOR)) return false;
-
-  var dim = X, val = Y, idx = 1;
-  if (encoding.isType(X,Q|T) && !encoding.isType(Y,Q|T) && encoding.has(Y)) {
-    dim = Y;
-    val = X;
-    idx = 0;
-  }
-
-  // add transform to compute sums for scale
-  var stacked = {
-    name: STACKED,
-    source: TABLE,
-    transform: [{
-      type: "aggregate",
-      groupby: [encoding.field(dim)].concat(facets), // dim and other facets
-      fields: [{op: "sum", field: encoding.field(val)}] // TODO check if field with aggr is correct?
-    }]
-  };
-
-  if(facets && facets.length > 0){
-    stacked.transform.push({ //calculate max for each facet
-      type: "aggregate",
-      groupby: facets,
-      fields: [{op: "max", field: "data.sum_" + encoding.field(val, true)}]
-    });
-  }
-
-  spec.data.push(stacked);
-
-  // add stack transform to mark
-  mdef.from.transform = [{
-    type: "stack",
-    point: encoding.field(dim),
-    height: encoding.field(val),
-    output: {y1: val, y0: val+"2"}
-  }];
-
-  // TODO: This is super hack-ish -- consolidate into modular mark properties?
-  mdef.properties.update[val] = mdef.properties.enter[val] = {scale: val, field: val};
-  mdef.properties.update[val+"2"] = mdef.properties.enter[val+"2"] = {scale: val, field: val+"2"};
-
-  return val; //return stack encoding
-}
-
-
-// BEGIN: AXES
-
-var axis = {};
-
-axis.names = function (props) {
-  return util.keys(util.keys(props).reduce(function(a, x) {
-    var s = props[x].scale;
-    if (s===X || s===Y) a[props[x].scale] = 1;
-    return a;
-  }, {}));
-}
-
-axis.defs = function(names, encoding, opt) {
-  return names.reduce(function(a, name) {
-    a.push(axis_def(name, encoding, opt));
-    return a;
-  }, []);
-}
-
-function axis_def(name, encoding, opt){
-  var type = name, axis;
-  var isCol = name==COL, isRow = name==ROW;
-  if(isCol) type = "x";
-  if(isRow) type = "y";
-
-  var axis = {
-    type: type,
-    scale: name,
-    ticks: 3 //TODO(kanitw): better determine # of ticks
-  };
-
-  if (encoding.axis(name).grid) {
-    axis.grid = true;
-    axis.layer = "back";
-  }
-
-  if (encoding.axis(name).title) {
-    axis.title = name;
-    // TODO: set appropriate titleOffset
-    // maybe based on some string length from stats
-  }
-
-  if(isRow || isCol){
-    axis.properties = {
-      ticks: { opacity: {value: 0} },
-      majorTicks: { opacity: {value: 0} },
-      axis: { opacity: {value: 0} }
-    };
-  }
-  if(isCol){
-    axis.offset = [opt.xAxisMargin || 0, encoding.config("yAxisMargin")];
-    axis.orient = "top";
-  }
-
-  if (name=="x" && (encoding.isType(name, O) || encoding.bin(name))) {
-    axis.properties = {
-      labels: {
-        angle: {value: 270},
-        align: {value: "right"},
-        baseline: {value: "middle"}
-      }
-    }
-  }
-
-  return axis;
-}
-
-// END: AXES
-
-// BEGIN: SCALE
-var scale = {};
-
-scale.names = function (props) {
-  return util.keys(util.keys(props).reduce(function(a, x) {
-    if (props[x] && props[x].scale) a[props[x].scale] = 1;
-    return a;
-  }, {}));
-}
-
-scale.defs = function (names, encoding, opt) {
-  opt = opt || {};
-
-  return names.reduce(function(a, name) {
-    var s = {
-      name: name,
-      type: scale_type(name, encoding),
-      domain: scale_domain(name, encoding, opt)
-    };
-    if (s.type === "ordinal" && !encoding.bin(name)) {
-      s.sort = true;
-    }
-
-    scale_range(s, encoding, opt);
-
-    return (a.push(s), a);
-  }, []);
-}
-
-function scale_type(name, encoding) {
-  switch (encoding.type(name)) {
-    case O: return "ordinal";
-    case T:
-      if (encoding.fn(name)) {
-        return "linear";
-      }
-      return "time";
-    case Q:
-      if (encoding.bin(name)) {
-        return "ordinal";
-      }
-      return encoding.scale(name).type || "linear";
-  }
-}
-
-function scale_domain(name, encoding, opt) {
-  if (encoding.type(name) === T){
-    switch(encoding.fn(name)){
-      case "second":
-      case "minute": return [0, 59];
-      case "hour": return [0, 23];
-      case "day": return [0, 6];
-      case "date": return [1, 31];
-      case "month": return [0, 11];
-    }
-  }
-
-  if (encoding.bin(name)) {
-    // TODO: add includeEmptyConfig here
-    if (opt.stats) {
-      var bins = util.getbins(opt.stats[encoding.fieldName(name)]);
-      var domain = util.range(bins.start, bins.stop, bins.step);
-      return name===Y ? domain.reverse() : domain;
-    }
-  }
-
-  return name == opt.stack ?
-    {
-      data: STACKED,
-      field: "data." + (opt.facet ? "max_" :"") + "sum_" + encoding.field(name, true)
-    }:
-    {data: TABLE, field: encoding.field(name)};
-}
-
-function scale_range(s, encoding, opt) {
-  var spec = encoding.scale(s.name);
-  switch (s.name) {
-    case X:
-      if (encoding.isType(s.name, O) || encoding.bin(s.name)) {
-        s.bandWidth = +encoding.config("bandSize");
-      } else {
-        s.range = opt.cellWidth ? [0, opt.cellWidth] : "width";
-        //TODO zero and reverse should become generic, and we just read default from either the schema or the schema generator
-        s.zero = spec.zero || encoding.config("_xZero");
-        s.reverse = spec.reverse || encoding.config("_xReverse");
-      }
-      s.round = true;
-      if (encoding.isType(s.name, T)){
-        s.nice = encoding.aggr(s.name) || encoding.config("timeScaleNice");
-      }else{
-        s.nice = true;
-      }
-      break;
-    case Y:
-      if (encoding.isType(s.name, O) || encoding.bin(s.name)) {
-        s.bandWidth = +encoding.config("bandSize");
-      } else {
-        s.range = opt.cellHeight ? [opt.cellHeight, 0] : "height";
-        //TODO zero and reverse should become generic, and we just read default from either the schema or the schema generator
-        s.zero = spec.zero || encoding.config("_yZero");
-        s.reverse = spec.reverse || encoding.config("_yReverse");
-      }
-
-      s.round = true;
-
-      if (encoding.isType(s.name, T)){
-        s.nice = encoding.aggr(s.name);
-      }else{
-        s.nice = true;
-      }
-      break;
-    case ROW:
-      s.bandWidth = opt.cellHeight || encoding.config("cellHeight");
-      s.round = true;
-      s.nice = true;
-      break;
-    case COL:
-      s.bandWidth = opt.cellWidth || encoding.config("cellWidth");
-      s.round = true;
-      s.nice = true;
-      break;
-    case SIZE:
-      if (encoding.is("bar")) {
-        s.range = [3, +encoding.config("bandSize")];
-      } else if (encoding.is(TEXT)) {
-        s.range = [8, 40];
-      } else {
-        s.range = [10, 1000];
-      }
-      s.round = true;
-      s.zero = false;
-      break;
-    case SHAPE:
-      s.range = "shapes";
-      break;
-    case COLOR:
-      if (encoding.isType(s.name, O)) {
-        s.range = "category10";
-      } else {
-        s.range = ["#ddf", "steelblue"];
-        s.zero = false;
-      }
-      break;
-    case ALPHA:
-      s.range = [0.2, 1.0];
-      break;
-    default:
-      throw new Error("Unknown encoding name: "+s.name);
-  }
-
-  switch(s.name){
-    case ROW:
-    case COL:
-      s.padding = encoding.config("cellPadding");
-      s.outerPadding = 0;
-      break;
-    case X:
-    case Y:
-      if (encoding.isType(s.name, O) || encoding.bin(s.name) ) { //&& !s.bandWidth
-        s.points = true;
-        s.padding = encoding.config("bandPadding");
-      }
-  }
-}
-
-// END: SCALE
-
-// BEGIN: MARKS
-function markdef(mark, encoding, opt) {
-  var p = mark.prop(encoding, opt)
-  return {
-    type: mark.type,
-    from: {data: TABLE},
-    properties: {enter: p, update: p}
-  };
-}
-
-function groupdef(name, opt) {
-  opt = opt || {};
-  return {
-    _name: name || undefined,
-    type: "group",
-    from: opt.from,
-    properties: {
-      enter: {
-        x: opt.x || undefined,
-        y: opt.y || undefined,
-        width: opt.width || {group: "width"},
-        height: opt.height || {group: "height"}
-      }
-    },
-    scales: opt.scales || undefined,
-    axes: opt.axes || undefined,
-    marks: opt.marks || []
-  };
-}
-
-function template(encoding, size, stats) { //hack use stats
-
-  var data = {name:TABLE, format: {type: encoding.config("dataFormatType")}},
-    dataUrl = vl.getDataUrl(encoding, stats);
-  if(dataUrl) data.url = dataUrl;
-
-  var preaggregatedData = encoding.config("useVegaServer");
-
-  encoding.forEach(function(encType, field){
-    if(field.type == T){
-      data.format.parse = data.format.parse || {};
-      data.format.parse[field.name] = "date";
-    }else if(field.type == Q){
-      data.format.parse = data.format.parse || {};
-      if (field.aggr === "count") {
-        var name = "count";
-      } else if(preaggregatedData && field.bin){
-        var name = "bin_" + field.name;
-      } else if(preaggregatedData && field.aggr){
-        var name = field.aggr + "_" + field.name;
-      } else{
-        var name = field.name;
-      }
-      data.format.parse[name] = "number";
-    }
-  });
-
-  return {
-    width: size.width,
-    height: size.height,
-    padding: "auto",
-    data: [data],
-    marks: [groupdef("cell", {
-      width: size.cellWidth ? {value: size.cellWidth}: undefined,
-      height: size.cellHeight ? {value: size.cellHeight} : undefined
-    })]
-  };
-}
-
-// --------------------------------------------------------
-
-var marks = vl.marks = {};
-
-marks.bar = {
-  type: "rect",
-  stack: true,
-  prop: bar_props,
-  requiredEncoding: ["x", "y"],
-  supportedEncoding: {row:1, col:1, x:1, y:1, size:1, color:1, alpha:1}
-};
-
-marks.line = {
-  type: "line",
-  line: true,
-  prop: line_props,
-  requiredEncoding: ["x", "y"],
-  supportedEncoding: {row:1, col:1, x:1, y:1, color:1, alpha:1}
-};
-
-marks.area = {
-  type: "area",
-  stack: true,
-  line: true,
-  requiredEncoding: ["x", "y"],
-  prop: area_props,
-  supportedEncoding: marks.line.supportedEncoding
-};
-
-marks.circle = {
-  type: "symbol",
-  prop: filled_point_props("circle"),
-  supportedEncoding: {row:1, col:1, x:1, y:1, size:1, color:1, alpha:1}
-};
-
-marks.square = {
-  type: "symbol",
-  prop: filled_point_props("square"),
-  supportedEncoding: marks.circle.supportedEncoding
-};
-
-marks.point = {
-  type: "symbol",
-  prop: point_props,
-  supportedEncoding: {row:1, col:1, x:1, y:1, size:1, color:1, alpha:1, shape:1}
-};
-
-marks.text = {
-  type: "text",
-  prop: text_props,
-  requiredEncoding: ["text"],
-  supportedEncoding: {row:1, col:1, size:1, color:1, alpha:1, text:1}
-};
-
-function bar_props(e) {
-  var p = {};
-
-  // x
-  if (e.isType(X,Q|T) && !e.bin(X)) {
-    p.x = {scale: X, field: e.field(X)};
-    if (e.has(Y) && (!e.isType(Y,Q|T) || e.bin(Y))) {
-      p.x2 = {scale: X, value: 0};
-    }
-  } else if (e.has(X)) {
-    p.xc = {scale: X, field: e.field(X)};
-  } else {
-    p.xc = {value: 0};
-  }
-
-  // y
-  if (e.isType(Y,Q|T) && !e.bin(Y)) {
-    p.y = {scale: Y, field: e.field(Y)};
-    p.y2 = {scale: Y, value: 0};
-  } else if (e.has(Y)) {
-    p.yc = {scale: Y, field: e.field(Y)};
-  } else {
-    p.yc = {group: "height"};
-  }
-
-  // width
-  if (!e.isType(X,Q|T)) {
-    if (e.has(SIZE)) {
-      p.width = {scale: SIZE, field: e.field(SIZE)};
-    } else {
-      // p.width = {scale: X, band: true, offset: -1};
-      p.width = {value: +e.config("bandSize"), offset: -1};
-    }
-  } else if (!e.isType(Y,O) && !e.bin(Y)) {
-    p.width = {value: +e.config("bandSize"), offset: -1};
-  }
-
-  // height
-  if (!e.isType(Y,Q|T)) {
-    if (e.has(SIZE)) {
-      p.height = {scale: SIZE, field: e.field(SIZE)};
-    } else {
-      // p.height = {scale: Y, band: true, offset: -1};
-      p.height = {value: +e.config("bandSize"), offset: -1};
-    }
-  } else if (!e.isType(X,O) && !e.bin(X)) {
-    p.height = {value: +e.config("bandSize"), offset: -1};
-  }
-
-  // fill
-  if (e.has(COLOR)) {
-    p.fill = {scale: COLOR, field: e.field(COLOR)};
-  } else if (!e.has(COLOR)) {
-    p.fill = {value: e.config("color")};
-  }
-
-  // alpha
-  if (e.has(ALPHA)) {
-    p.opacity = {scale: ALPHA, field: e.field(ALPHA)};
-  }
-
-  return p;
-}
-
-function point_props(e, opt) {
-  var p = {};
-  opt = opt || {};
-
-  // x
-  if (e.has(X)) {
-    p.x = {scale: X, field: e.field(X)};
-  } else if (!e.has(X)) {
-    p.x = {value: e.config("bandSize")/2};
-  }
-
-  // y
-  if (e.has(Y)) {
-    p.y = {scale: Y, field: e.field(Y)};
-  } else if (!e.has(Y)) {
-    p.y = {value: e.config("bandSize")/2};
-  }
-
-  // size
-  if (e.has(SIZE)) {
-    p.size = {scale: SIZE, field: e.field(SIZE)};
-  } else if (!e.has(SIZE)) {
-    p.size = {value: e.config("pointSize")};
-  }
-
-  // shape
-  if (e.has(SHAPE)) {
-    p.shape = {scale: SHAPE, field: e.field(SHAPE)};
-  } else if (!e.has(SHAPE)) {
-    p.shape = {value: e.config("pointShape")};
-  }
-
-  // stroke
-  if (e.has(COLOR)) {
-    p.stroke = {scale: COLOR, field: e.field(COLOR)};
-  } else if (!e.has(COLOR)) {
-    p.stroke = {value: e.config("color")};
-  }
-
-  // alpha
-  if (e.has(ALPHA)) {
-    p.opacity = {scale: ALPHA, field: e.field(ALPHA)};
-  }else{
-    p.opacity = {
-      value: e.config("opacity") || e.config(opt.hasAggregate ? "_thickOpacity" : "_thinOpacity")
-    };
-  }
-
-  p.strokeWidth = {value: e.config("strokeWidth")};
-
-  return p;
-}
-
-function line_props(e) {
-  var p = {};
-
-  // x
-  if (e.has(X)) {
-    p.x = {scale: X, field: e.field(X)};
-  } else if (!e.has(X)) {
-    p.x = {value: 0};
-  }
-
-  // y
-  if (e.has(Y)) {
-    p.y = {scale: Y, field: e.field(Y)};
-  } else if (!e.has(Y)) {
-    p.y = {group: "height"};
-  }
-
-  // stroke
-  if (e.has(COLOR)) {
-    p.stroke = {scale: COLOR, field: e.field(COLOR)};
-  } else if (!e.has(COLOR)) {
-    p.stroke = {value: e.config("color")};
-  }
-
-  // alpha
-  if (e.has(ALPHA)) {
-    p.opacity = {scale: ALPHA, field: e.field(ALPHA)};
-  }
-
-  p.strokeWidth = {value: e.config("strokeWidth")};
-
-  return p;
-}
-
-function area_props(e) {
-  var p = {};
-
-  // x
-  if (e.isType(X,Q|T)) {
-    p.x = {scale: X, field: e.field(X)};
-    if (!e.isType(Y,Q|T) && e.has(Y)) {
-      p.x2 = {scale: X, value: 0};
-      p.orient = {value: "horizontal"};
-    }
-  } else if (e.has(X)) {
-    p.x = {scale: X, field: e.field(X)};
-  } else {
-    p.x = {value: 0};
-  }
-
-  // y
-  if (e.isType(Y,Q|T)) {
-    p.y = {scale: Y, field: e.field(Y)};
-    p.y2 = {scale: Y, value: 0};
-  } else if (e.has(Y)) {
-    p.y = {scale: Y, field: e.field(Y)};
-  } else {
-    p.y = {group: "height"};
-  }
-
-  // stroke
-  if (e.has(COLOR)) {
-    p.fill = {scale: COLOR, field: e.field(COLOR)};
-  } else if (!e.has(COLOR)) {
-    p.fill = {value: e.config("color")};
-  }
-
-  // alpha
-  if (e.has(ALPHA)) {
-    p.opacity = {scale: ALPHA, field: e.field(ALPHA)};
-  }
-
-  return p;
-}
-
-function filled_point_props(shape) {
-  return function(e, opt) {
-    var p = {};
-    opt = opt || {};
-
-    // x
-    if (e.has(X)) {
-      p.x = {scale: X, field: e.field(X)};
-    } else if (!e.has(X)) {
-      p.x = {value: e.config("bandSize")/2};
-    }
-
-    // y
-    if (e.has(Y)) {
-      p.y = {scale: Y, field: e.field(Y)};
-    } else if (!e.has(Y)) {
-      p.y = {value: e.config("bandSize")/2};
-    }
-
-    // size
-    if (e.has(SIZE)) {
-      p.size = {scale: SIZE, field: e.field(SIZE)};
-    } else if (!e.has(X)) {
-      p.size = {value: e.config("pointSize")};
-    }
-
-    // shape
-    p.shape = {value: shape};
-
-    // fill
-    if (e.has(COLOR)) {
-      p.fill = {scale: COLOR, field: e.field(COLOR)};
-    } else if (!e.has(COLOR)) {
-      p.fill = {value: e.config("color")};
-    }
-
-    // alpha
-    if (e.has(ALPHA)) {
-      p.opacity = {scale: ALPHA, field: e.field(ALPHA)};
-    }else {
-      p.opacity = {
-        value: e.config("opacity") || e.config(opt.hasAggregate ? "_thickOpacity" : "_thinOpacity")
-      };
-    }
-
-    return p;
-  };
-}
-
-function text_props(e) {
-  var p = {};
-
-  // x
-  if (e.has(X)) {
-    p.x = {scale: X, field: e.field(X)};
-  } else if (!e.has(X)) {
-    p.x = {value: e.config("bandSize")/2};
-  }
-
-  // y
-  if (e.has(Y)) {
-    p.y = {scale: Y, field: e.field(Y)};
-  } else if (!e.has(Y)) {
-    p.y = {value: e.config("bandSize")/2};
-  }
-
-  // size
-  if (e.has(SIZE)) {
-    p.fontSize = {scale: SIZE, field: e.field(SIZE)};
-  } else if (!e.has(X)) {
-    p.fontSize = {value: e.config("fontSize")};
-  }
-
-  // fill
-  if (e.has(COLOR)) {
-    p.fill = {scale: COLOR, field: e.field(COLOR)};
-  } else if (!e.has(COLOR)) {
-    p.fill = {value: e.config("textColor")};
-  }
-
-  // alpha
-  if (e.has(ALPHA)) {
-    p.opacity = {scale: ALPHA, field: e.field(ALPHA)};
-  }
-
-  // text
-  if (e.has(TEXT)) {
-    p.text = {field: e.field(TEXT)};
-  } else {
-    p.text = {value: "Abc"};
-  }
-
-  p.font = {value: e.config("font")};
-  p.fontWeight = {value: e.config("fontWeight")};
-  p.fontStyle = {value: e.config("fontStyle")};
-  p.baseline = {value: e.config("textBaseline")};
-
-  // align
-  if (e.has(X)) {
-    if (e.isType(X,O)) {
-      p.align = {value: "left"};
-      p.dx = {value: e.config("textMargin")};
-    } else {
-      p.align = {value: "center"}
-    }
-  } else if (e.has(Y)) {
-    p.align = {value: "left"};
-    p.dx = {value: e.config("textMargin")};
-  } else {
-    p.align = {value: e.config("textAlign")};
-  }
-
-  return p;
-}
-
-// END MARKS
-
-// BEGIN LEGENDS
-
-var legends = {};
-legends.defs = function(encoding) {
-  var legends = [];
-
-  // TODO: support alpha
-
-  if (encoding.has("color") && encoding.legend("color")) {
-    legends.push({
-      fill: "color",
-      title: encoding.fieldName("color"),
-      orient: "right"
-    });
-  }
-
-  if (encoding.has("size") && encoding.legend("size")) {
-    legends.push({
-      size: "size",
-      title: encoding.fieldName("size"),
-      orient: legends.length === 1 ? "left" : "right"
-    });
-  }
-
-  if (encoding.has("shape") && encoding.legend("shape")) {
-    if (legends.length === 2) {
-      // TODO: fix this
-      console.error("Vegalite currently only supports two legends");
-      return legends;
-    }
-    legends.push({
-      shape: "shape",
-      title: encoding.fieldName("shape"),
-      orient: legends.length === 1 ? "left" : "right"
-    });
-  }
-
-  return legends;
-}
-
-// END LEGENDS
-
-return vl;
-// END OF THIS MODULE
-}));
